pyproject.toml
[project]
name = "fastapi-hexagonal-todo"
version = "0.1.0"
description = "FastAPI Todo application with Hexagonal Architecture"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Developer", email = "dev@example.com"}
]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic[email]>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.23",
    "aiosqlite>=0.19.0",
    "alembic>=1.13.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
dev = "app.cli:run_dev"
start = "app.cli:run_start"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
